gen-ai-sandbox/src/chatbot_app/requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
retrying==1.3.4
llama-index-core==0.10.57
google-cloud-firestore==2.16.1
langchain-google-vertexai==0.0.5
llama-index-embeddings-gemini==0.1.4
llama-index-embeddings-huggingface==0.2.0
llama-index-llms-vertex==0.2.2
llama-index-llms-gemini==0.1.4
llama-index-llms-huggingface==0.2.0
llama-index-llms-langchain==0.1.3
llama-index-readers-file==0.1.30
llama-index-vector-stores-pinecone==0.1.8
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pinecone-client==3.2.2
wget==3.2
streamlit==1.38.0
